Create grott_ha_spf.py
Changed out the mapping completely to what I have mapped to a an SPF5000ES. It works the same for the SPF3500ES and possibly most SPF inverters. I just rename this file to grott_ha.py and put it into the root directory of the grott zip file for installation on my Home Assistant using Docker.